build: reject implicit engine size narrowing on MSVC

+++ engine/botlib/be_aas_bspq3.cpp
@@ -481,7 +481,7 @@ void AAS_DumpBSPData(void)
 int AAS_LoadBSPFile(void)
 {
 	AAS_DumpBSPData();
-	bspworld.entdatasize = strlen(botimport.BSPEntityData()) + 1;
+	bspworld.entdatasize = (int)( strlen(botimport.BSPEntityData()) + 1 );
 	bspworld.dentdata = (char *) GetClearedHunkMemory(bspworld.entdatasize);
 	Com_Memcpy(bspworld.dentdata, botimport.BSPEntityData(), bspworld.entdatasize);
 	AAS_ParseBSPEntities();
