+++ engine/botlib/l_struct.cpp
@@ -332,7 +332,7 @@ int WriteFloat(FILE *fp, float value)
 	int l;
 
 	Com_sprintf(buf, sizeof(buf), "%f", value);
-	l = strlen(buf);
+	l = (int)( strlen(buf) );
 	//strip any trailing zeros
 	while(l-- > 1)
 	{
