+++ engine/qcommon/q_shared.cpp
@@ -99,8 +99,8 @@ qboolean COM_CompareExtension(const char *in, const char *ext)
 {
 	int inlen, extlen;
 	
-	inlen = strlen(in);
-	extlen = strlen(ext);
+	inlen = (int)( strlen(in) );
+	extlen = (int)( strlen(ext) );
 	
 	if(extlen <= inlen)
 	{
@@ -1073,7 +1073,7 @@ int Com_HexStrToInt( const char *str )
 	// check for hex code
 	if ( str[ 0 ] == '0' && str[ 1 ] == 'x' && str[ 2 ] != '\0' )
 	{
-		int i, digit, n = 0, len = strlen( str );
+		int i, digit, n = 0, len = (int)strlen( str );
 
 		for( i = 2; i < len; i++ )
 		{
@@ -1260,7 +1260,7 @@ int Q_vsnprintf( char *str, size_t size, const char *format, va_list ap )
 		// implementation, so we have no choice but to return size.
 		
 		str[size - 1] = '\0';
-		return size;
+		return (int)( size );
 	}
 	
 	return retval;
@@ -1506,7 +1506,7 @@ char *Q_strupr( char *s1 ) {
 void Q_strcat( char *dest, int size, const char *src ) {
 	int		l1;
 
-	l1 = strlen( dest );
+	l1 = (int)( strlen( dest ) );
 	if ( l1 >= size ) {
 		Com_Error( ERR_FATAL, "Q_strcat: already overflowed" );
 	}
@@ -1550,7 +1550,7 @@ const char *Q_stristr( const char *s, const char *find)
           sc -= ('a' - 'A');
         }
       } while (sc != c);
-    } while (Q_stricmpn(s, find, len) != 0);
+    } while (Q_stricmpn(s, find, (int)( len )) != 0);
     s--;
   }
   return s;
@@ -1570,8 +1570,8 @@ int Q_replace( const char *str1, const char *str2, char *src, int max_len )
 
 	count = 0; // replace count
 
-    len1 = strlen( str1 );
-    len2 = strlen( str2 );
+    len1 = (int)( strlen( str1 ) );
+    len2 = (int)( strlen( str2 ) );
     d = len2 - len1;
 
     if ( d > 0 ) // expand and replace mode    
@@ -1792,7 +1792,7 @@ Assumes buffer is at least TRUNCATE_LENGTH big
 */
 void Com_TruncateLongString( char *buffer, const char *s )
 {
-	int length = strlen( s );
+	int length = (int)( strlen( s ) );
 
 	if( length <= TRUNCATE_LENGTH )
 		Q_strncpyz( buffer, s, TRUNCATE_LENGTH );
